requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-celery-results==2.5.1
django-redis==5.4.0
mysqlclient==2.2.7
python-dotenv==1.0.1
celery==5.4.0
redis==5.2.1
chromadb==0.5.23
sentence-transformers==3.3.1
openai==1.58.1
google-genai==1.33.0
selenium==4.27.1
webdriver-manager==4.0.2
requests==2.32.3
beautifulsoup4==4.12.3
gunicorn==23.0.0
